Combine Select with unwriting operations
